% Auto-generated: do not edit by hand
\name{''DashFlow}

\alias{''DashFlow}

\title{DashFlow component}

\description{

}

\usage{
''DashFlow(id=NULL, elements=NULL, fitView=NULL, layout=NULL,
onConnect=NULL, onEdgesChange=NULL, onElementsRemove=NULL,
onNodeClick=NULL, onNodeDragStop=NULL, onNodesChange=NULL,
onPaneClick=NULL)
}

\arguments{
\item{id}{Character. }

\item{elements}{Named list. }

\item{fitView}{Logical. }

\item{layout}{Logical. }

\item{onConnect}{Named list. }

\item{onEdgesChange}{Unnamed list. }

\item{onElementsRemove}{Unnamed list. }

\item{onNodeClick}{Character. }

\item{onNodeDragStop}{Named list. }

\item{onNodesChange}{Unnamed list. }

\item{onPaneClick}{Lists containing elements 'x', 'y'.
those elements have the following types:
  - x (numeric; optional)
  - y (numeric; optional). }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

